Clutter -- create helper_functions
